Working Scraper and URL Results
To avoid scraping without permission a few scrape tests where done and results saved to allow for testing and building after the scrape step.

The Thoughts card is output before the generated search URL's to see if the LLM's logic is good.  This might be a good spot to add a button to continue to the scraping step, so the user can redo the LLM quarry without triggering a scrape each time.

Currently the lists of results URL's for each search are placed in accordian_pannels.  These may or may not stay, but its the links we need to build the next step of Abstract summaries.