Working with Search Scraping Bypassed
This is a fully working version.  The search results scraping is currently bypassed as we determine if its ok to scrape search results.  Using a preset search results file, the rest of the app has been built and is working.

The app flow is this:
Search for something mining related, and get a list of suggested search URLS.  They can be clicked on to see search results.

The user is presented with a button to scrape the search results.  When clicked the user is taken to the Scraped tab, and shown the articles pulled from the scraping of search links.  After reviewing the user can click a "Summarize Search Results" button.

The Summarize button will bring the user to the results page.  Which gives a summary and final selection of the most useful articles from the search results for the users original search.  It also lists the articles, and includes links directly to them.